requirements.txt
setuptools
wheel
fastapi
uvicorn
httpx
beautifulsoup4
scrapy
apify-client
cachetools
requests
nltk
matplotlib
wordcloud
pandas
python-dotenv
pydantic
numpy
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-groq
langchain-openai
langchain-text-splitters
langgraph
langsmith
arxiv